scripts/requirements.txt
requests
websocket-client
markdownify
torch --index-url https://download.pytorch.org/whl/cpu
